requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
django-dotenv==1.4.2
gunicorn==22.0.0
packaging==24.0
psycopg2==2.9.9
PyMySQL==1.1.1
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
whitenoise==6.6.0
